Store the angle rotation directly instead of the packed stoneId
