+++ contracts/contracts/interfaces/IWands.sol
@@ -21,7 +21,7 @@ interface IWands is IERC721 {
 
   struct Wand {
     bool built;
-    uint32 stone;
+    uint16 stone;
     uint8 handle;
     uint16 halo;
     uint32 evolution;
@@ -33,7 +33,7 @@ interface IWands is IERC721 {
 
   function build(
     uint256 tokenId,
-    uint32 stone,
+    uint16 stone,
     uint8 handle,
     uint16 halo,
     Template.Background memory background,
